app/backend/requirements.txt
asttokens
backcall
beautifulsoup4
blinker
cachetools
certifi
cffi
chardet
click
cryptography
decorator
executing
flask
Flask-Cors
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
idna
importlib-metadata
ipython
itsdangerous
jedi
jinja2
MarkupSafe
matplotlib-inline
parso
pdfminer.six
pdfx
pexpect
pickleshare
prompt-toolkit
proto-plus
protobuf
ptyprocess
pure-eval
pyasn1
pyasn1-modules
pycparser
pygments
requests
rsa
six
sortedcontainers
soupsieve
stack-data
traitlets
typing-extensions
urllib3
wcwidth
werkzeug
zipp
